+++ components/merkledb/src/error.rs
@@ -16,12 +16,11 @@
 
 use std::error::Error as StdError;
 
-/// The error type for I/O operations with storage.
+/// The error type for I/O operations with the `Database`.
 ///
-/// These errors result in a panic. Storage errors are fatal as in the case of
-/// database issues, the system stops working. Assuming that there are other
-/// nodes and secret keys and other crucial data are not stored in the data base,
-/// the operation of the system can be resumed from a backup or by rebooting the node.
+/// Application code in most cases should consider these errors as fatal. At the same time,
+/// it may be possible to recover from an error after manual intervention (e.g., by restarting
+/// the process or freeing up more disc space).
 #[derive(Fail, Debug, Clone)]
 #[fail(display = "{}", message)]
 pub struct Error {