+++ components/merkledb/src/hash.rs
@@ -21,49 +21,52 @@ use crate::{indexes::proof_map_index::ProofPath, BinaryValue};
 const EMPTY_LIST_HASH: &str = "c6c0aa07f27493d2f2e5cff56c890a353a20086d6c25ec825128e12ae752b2d9";
 const EMPTY_MAP_HASH: &str = "7324b5c72b51bb5d4c180f1109cfd347b60473882145841c39f3e584576296f9";
 
-#[repr(u8)]
-#[derive(Copy, Clone, Debug)]
-/// `MerkleDB` hash prefixes.
-pub enum HashTag {
-    /// Hash prefix of a blob.
-    Blob = 0,
-    /// Hash prefix of a branch node of the merkle tree.
-    ListBranchNode = 1,
-    /// Hash prefix of the list object.
-    ListNode = 2,
-    /// Hash prefix of the map object.
-    MapNode = 3,
-    /// Hash prefix of the map branch node object.
-    MapBranchNode = 4,
-}
-
-/// Calculate hash value with the specified prefix.
+/// Prefixes for different types of objects stored in the database. These prefixes are necessary
+/// to provide domain separation among hashed objects of different types.
 ///
 /// In `MerkleDB`, all data is presented as objects. Objects are divided into blobs
 /// and collections (lists / maps), which in their turn are divided into hashable and
-/// non-hashable. `ProofListIndex` and `ProofMapIndex` relate to hashable collections.
-/// For these collections, one can define a hash, which is used to build proof for
-/// their contents. In the future, these hashes will be used to build proofs for object
-/// hierarchies.
+/// non-hashable. The hashable collections are `ProofListIndex` and `ProofMapIndex`.
+/// For these collections, one can define a single hash which would reflect the entire
+/// collection contents. This hash can then be used that a collection contains (or does not contain)
+/// certain elements.
 ///
 /// Different hashes for leaf and branch nodes of the list are used to secure merkle tree
-/// from the pre-image attack.
+/// from the pre-image attack. See more information [here][rfc6962].
 ///
-/// See more information [here][1].
-///
-/// [1]: https://tools.ietf.org/html/rfc6962#section-2.1
+/// [rfc6962]: https://tools.ietf.org/html/rfc6962#section-2.1
+#[derive(Copy, Clone, Debug)]
+#[repr(u8)]
+pub enum HashTag {
+    /// Hash prefix of a blob (i.e., a type implementing [`BinaryValue`], which is stored in the DB
+    /// as byte sequence).
+    ///
+    /// [`BinaryValue`]: trait.BinaryValue.html
+    Blob = 0,
+    /// Hash prefix of a branch node in a Merkle tree built for
+    /// a [Merkelized list](proof_list_index/struct.ProofListIndex.html).
+    ListBranchNode = 1,
+    /// Hash prefix of a [Merkelized list](proof_list_index/struct.ProofListIndex.html).
+    ListNode = 2,
+    /// Hash prefix of a [Merkelized map](proof_map_index/struct.ProofMapIndex.html).
+    MapNode = 3,
+    /// Hash prefix of a branch node in a Merkle Patricia tree built for
+    /// a [Merkelized map](proof_map_index/struct.ProofMapIndex.html).
+    MapBranchNode = 4,
+}
+
 impl HashTag {
     ///`HashStream` object with the corresponding hash prefix.
     pub(crate) fn hash_stream(self) -> HashStream {
         HashStream::new().update(&[self as u8])
     }
 
-    /// Convenience method to obtain a hashed value of the merkle tree leaf.
+    /// Obtains a hashed value of a leaf in a Merkle tree.
     pub fn hash_leaf(value: &[u8]) -> Hash {
         HashTag::Blob.hash_stream().update(value).hash()
     }
 
-    /// Convenience method to obtain hashed value of the merkle tree node.
+    /// Obtains a hashed value of a branch in a Merkle tree.
     pub fn hash_node(left_hash: &Hash, right_hash: &Hash) -> Hash {
         HashTag::ListBranchNode
             .hash_stream()
@@ -72,7 +75,7 @@ impl HashTag {
             .hash()
     }
 
-    /// Convenience method to obtain a hashed value of the merkle tree node with one child.
+    /// Obtains a hashed value of a Merkle tree branch with one child.
     pub fn hash_single_node(hash: &Hash) -> Hash {
         HashTag::ListBranchNode
             .hash_stream()
@@ -80,7 +83,8 @@ impl HashTag {
             .hash()
     }
 
-    /// Hash of the list object.
+    /// Obtains hash of a Merkelized list. `len` is the length of the list, and `root` is
+    /// the hash of the root node of the Merkle tree corresponding to the list.
     ///
     /// ```text
     /// h = sha-256( HashTag::ListNode || len as u64 || merkle_root )
@@ -96,22 +100,22 @@ impl HashTag {
             .hash()
     }
 
-    /// Hash of the empty list object.
+    /// Hash of an empty Merkelized list.
     ///
-    /// Empty list hash:
     /// ```text
-    /// h = sha-256( HashTag::ListNode || 0 || Hash::default() )
+    /// h = sha-256( HashTag::ListNode || 0 || Hash::zero() )
     /// ```
     pub fn empty_list_hash() -> Hash {
         Hash::from_hex(EMPTY_LIST_HASH).unwrap()
     }
 
-    /// Computes a list hash for the given list of hashes.
+    /// Computes the hash for a Merkelized list containing the given values.
     pub fn hash_list<V: BinaryValue + ?Sized>(values: &[V]) -> Hash {
         Self::hash_list_node(values.len() as u64, root_hash(values))
     }
 
-    /// Hash of the map object.
+    /// Hash of a Merkelized map with at least 2 entries. `root` is the recursively defined
+    /// hash of the root node of the binary Patricia Merkle tree corresponding to the map.
     ///
     /// ```text
     /// h = sha-256( HashTag::MapNode || merkle_root )
@@ -123,11 +127,16 @@ impl HashTag {
             .hash()
     }
 
-    /// Hash of the map branch node.
+    /// Hash of a branch node in a Merkle Patricia tree. `branch_node` is the binary serialization
+    /// of the node.
     ///
     /// ```text
-    /// h = sha-256( HashTag::MapBranchNode || <left_key> || <right_key> || <left_hash> || <right_hash> )
+    /// h = sha-256(HashTag::MapBranchNode || <branch_node>)
     /// ```
+    ///
+    /// See [`ProofMapIndex`] for details how branch nodes are serialized.
+    ///
+    /// [`ProofMapIndex`]: proof_map_index/struct.ProofMapIndex.html#impl-ObjectHash
     pub fn hash_map_branch(branch_node: &[u8]) -> Hash {
         HashStream::new()
             .update(&[HashTag::MapBranchNode as u8])
@@ -135,20 +144,20 @@ impl HashTag {
             .hash()
     }
 
-    /// Hash of the map with single entry.
+    /// Hash of a Merkelized map with a single entry.
     ///
     /// ``` text
-    /// h = sha-256( HashTag::MapBranchNode || <key> || <child_hash> )
+    /// h = sha-256( HashTag::MapBranchNode || <path> || <child_hash> )
     /// ```
-    pub fn hash_single_entry_map(path: &ProofPath, h: &Hash) -> Hash {
+    pub fn hash_single_entry_map(path: &ProofPath, child_hash: &Hash) -> Hash {
         HashStream::new()
             .update(&[HashTag::MapBranchNode as u8])
             .update(path.as_bytes())
-            .update(h.as_ref())
+            .update(child_hash.as_ref())
             .hash()
     }
 
-    /// Hash of the empty map object.
+    /// Hash of an empty Merkelized map.
     ///
     /// Empty map hash:
     /// ```text
@@ -199,9 +208,9 @@ pub fn root_hash<V: BinaryValue + ?Sized>(hashes: &[V]) -> Hash {
 
 /// A common trait for the ability to compute a unique hash.
 ///
-/// Unlike `CryptoHash`, the hash value returned by the `ObjectHash::hash()`
-/// method isn't always irreversible. This hash is used, for example, in the
-/// storage as a key, as uniqueness is important in this case.
+/// The hash value returned by the `object_hash()` method isn't always irreversible.
+/// This hash is used, for example, in the storage as a key, as uniqueness is important
+/// in this case.
 pub trait ObjectHash {
     /// Returns a hash of the value.
     ///
@@ -209,7 +218,7 @@ pub trait ObjectHash {
     fn object_hash(&self) -> Hash;
 }
 
-/// Just returns the origin hash.
+/// Just returns the original hash.
 impl ObjectHash for Hash {
     fn object_hash(&self) -> Hash {
         *self