+++ exonum/src/api/backends/actix.rs
@@ -21,10 +21,11 @@ pub use actix_web::middleware::cors::Cors;
 
 use actix::{Actor, System};
 use actix_web::{
-    error::ResponseError, AsyncResponder, FromRequest, HttpMessage, HttpResponse, Query,
+    error::ResponseError, http::header, AsyncResponder, FromRequest, HttpMessage, HttpResponse,
+    Query,
 };
 use failure::Error;
-use futures::{sync::mpsc, Future, IntoFuture, Stream};
+use futures::{future::Either, sync::mpsc, Future, IntoFuture, Stream};
 use serde::{
     de::{self, DeserializeOwned},
     ser, Serialize,
@@ -42,8 +43,8 @@ use std::{
 use crate::api::{
     self,
     manager::{ApiManager, UpdateEndpoints},
-    ApiAccess, ApiAggregator, ApiBackend, ApiScope, ExtendApiBackend, FutureResult, Immutable,
-    Mutable, NamedWith,
+    Actuality, ApiAccess, ApiAggregator, ApiBackend, ApiScope, EndpointMutability,
+    ExtendApiBackend, FutureResult, NamedWith,
 };
 
 /// Type alias for the concrete `actix-web` HTTP response.
@@ -131,108 +132,141 @@ impl ResponseError for api::Error {
             }
             api::Error::Io(err) => HttpResponse::InternalServerError().body(err.to_string()),
             api::Error::Storage(err) => HttpResponse::InternalServerError().body(err.to_string()),
+            api::Error::Gone => HttpResponse::Gone().finish(),
+            api::Error::MovedPermanently(new_location) => HttpResponse::MovedPermanently()
+                .header(header::LOCATION, new_location.clone())
+                .finish(),
             api::Error::NotFound(err) => HttpResponse::NotFound().body(err.to_string()),
             api::Error::Unauthorized => HttpResponse::Unauthorized().finish(),
         }
     }
 }
 
-impl<Q, I, F> From<NamedWith<Q, I, api::Result<I>, F, Immutable>> for RequestHandler
+/// Creates a `HttpResponse` object from the provided JSON value.
+/// Depending on the `actuality` parameter value, the warning about endpoint
+/// being deprecated can be added.
+fn json_response<T: Serialize>(actuality: Actuality, json_value: T) -> HttpResponse {
+    let mut response = HttpResponse::Ok();
+
+    if let Actuality::Deprecated {
+        ref discontinued_on,
+        ref description,
+    } = actuality
+    {
+        // There is a proposal for creating special deprecation header within HTTP,
+        // but currently it's only a draft. So the conventional way to notify API user
+        // about endpoint deprecation is setting the `Warning` header.
+        let expiration_note = match discontinued_on {
+            // Date is formatted according to HTTP-date format.
+            Some(date) => format!(
+                "The old API is maintained until {}.",
+                date.format("%a, %d %b %Y %T GMT")
+            ),
+            None => "Currently there is no specific date for disabling this endpoint.".into(),
+        };
+
+        let mut warning_text = format!(
+            "Deprecated API: This endpoint is deprecated, \
+             see the service documentation to find an alternative. \
+             {}",
+            expiration_note
+        );
+
+        if let Some(description) = description {
+            warning_text = format!("{} Additional information: {}.", warning_text, description);
+        }
+
+        let warning_string = create_warning_header(&warning_text);
+
+        response.header(header::WARNING, warning_string);
+    }
+
+    response.json(json_value)
+}
+
+/// Formats warning string according to the following format:
+/// "<warn-code> <warn-agent> \"<warn-text>\" [<warn-date>]"
+/// <warn-code> in our case is 299, which means a miscellaneous persistent warning.
+/// <warn-agent> is optional, so we set it to "-".
+/// <warn-text> is a warning description, which is taken as an only argument.
+/// <warn-date> is not required.
+/// For details you can see RFC 7234, section 5.5: Warning.
+fn create_warning_header(warning_text: &str) -> String {
+    format!("299 - \"{}\"", warning_text)
+}
+
+impl From<EndpointMutability> for actix_web::http::Method {
+    fn from(mutability: EndpointMutability) -> Self {
+        match mutability {
+            EndpointMutability::Immutable => actix_web::http::Method::GET,
+            EndpointMutability::Mutable => actix_web::http::Method::POST,
+        }
+    }
+}
+
+impl<Q, I, F> From<NamedWith<Q, I, api::Result<I>, F>> for RequestHandler
 where
     F: Fn(Q) -> api::Result<I> + 'static + Send + Sync + Clone,
     Q: DeserializeOwned + 'static,
     I: Serialize + 'static,
 {
-    fn from(f: NamedWith<Q, I, api::Result<I>, F, Immutable>) -> Self {
+    fn from(f: NamedWith<Q, I, api::Result<I>, F>) -> Self {
+        // Convert handler that returns a `Result` into handler that will return `FutureResult`.
         let handler = f.inner.handler;
-        let index = move |request: HttpRequest| -> FutureResponse {
-            let future = Query::from_request(&request, &Default::default())
-                .map(Query::into_inner)
-                .and_then(|query| handler(query).map_err(From::from))
-                .and_then(|value| Ok(HttpResponse::Ok().json(value)))
-                .into_future();
+        let future_endpoint = move |query| -> Box<dyn Future<Item = I, Error = api::Error>> {
+            let future = handler(query).into_future();
             Box::new(future)
         };
+        let named_with_future = NamedWith::new(f.name, future_endpoint, f.mutability);
 
-        Self {
-            name: f.name,
-            method: actix_web::http::Method::GET,
-            inner: Arc::from(index) as Arc<RawHandler>,
-        }
+        // Then we can create a `RequestHandler` with the `From` specialization for future result.
+        RequestHandler::from(named_with_future)
     }
 }
 
-impl<Q, I, F> From<NamedWith<Q, I, api::Result<I>, F, Mutable>> for RequestHandler
+/// Takes `HttpRequest` as a parameter and extracts query:
+/// - If request is immutable, the query is parsed from query string,
+/// - If request is mutable, the query is parsed from the request body as JSON.
+fn extract_query<Q>(
+    request: HttpRequest,
+    mutability: EndpointMutability,
+) -> impl Future<Item = Q, Error = actix_web::error::Error>
 where
-    F: Fn(Q) -> api::Result<I> + 'static + Send + Sync + Clone,
     Q: DeserializeOwned + 'static,
-    I: Serialize + 'static,
 {
-    fn from(f: NamedWith<Q, I, api::Result<I>, F, Mutable>) -> Self {
-        let handler = f.inner.handler;
-        let index = move |request: HttpRequest| -> FutureResponse {
-            let handler = handler.clone();
-            request
-                .json()
-                .from_err()
-                .and_then(move |query: Q| {
-                    handler(query)
-                        .map(|value| HttpResponse::Ok().json(value))
-                        .map_err(From::from)
-                })
-                .responder()
-        };
-
-        Self {
-            name: f.name,
-            method: actix_web::http::Method::POST,
-            inner: Arc::from(index) as Arc<RawHandler>,
-        }
-    }
-}
-
-impl<Q, I, F> From<NamedWith<Q, I, FutureResult<I>, F, Immutable>> for RequestHandler
-where
-    F: Fn(Q) -> FutureResult<I> + 'static + Clone + Send + Sync,
-    Q: DeserializeOwned + 'static,
-    I: Serialize + 'static,
-{
-    fn from(f: NamedWith<Q, I, FutureResult<I>, F, Immutable>) -> Self {
-        let handler = f.inner.handler;
-        let index = move |request: HttpRequest| -> FutureResponse {
-            let handler = handler.clone();
-            Query::from_request(&request, &Default::default())
+    match mutability {
+        EndpointMutability::Immutable => {
+            let future = Query::from_request(&request, &Default::default())
                 .map(Query::into_inner)
-                .into_future()
-                .and_then(move |query| handler(query).map_err(From::from))
-                .map(|value| HttpResponse::Ok().json(value))
-                .responder()
-        };
+                .map_err(From::from)
+                .into_future();
 
-        Self {
-            name: f.name,
-            method: actix_web::http::Method::GET,
-            inner: Arc::from(index) as Arc<RawHandler>,
+            Either::A(future)
+        }
+        EndpointMutability::Mutable => {
+            let future = request.json().from_err();
+            Either::B(future)
         }
     }
 }
 
-impl<Q, I, F> From<NamedWith<Q, I, FutureResult<I>, F, Mutable>> for RequestHandler
+impl<Q, I, F> From<NamedWith<Q, I, FutureResult<I>, F>> for RequestHandler
 where
     F: Fn(Q) -> FutureResult<I> + 'static + Clone + Send + Sync,
     Q: DeserializeOwned + 'static,
     I: Serialize + 'static,
 {
-    fn from(f: NamedWith<Q, I, FutureResult<I>, F, Mutable>) -> Self {
+    fn from(f: NamedWith<Q, I, FutureResult<I>, F>) -> Self {
         let handler = f.inner.handler;
+        let actuality = f.inner.actuality;
+        let mutability = f.mutability;
         let index = move |request: HttpRequest| -> FutureResponse {
             let handler = handler.clone();
-            request
-                .json()
-                .from_err()
-                .and_then(move |query: Q| {
+            let actuality = actuality.clone();
+            extract_query(request, mutability)
+                .and_then(move |query| {
                     handler(query)
-                        .map(|value| HttpResponse::Ok().json(value))
+                        .map(|value| json_response(actuality, value))
                         .map_err(From::from)
                 })
                 .responder()
@@ -240,7 +274,7 @@ where
 
         Self {
             name: f.name,
-            method: actix_web::http::Method::POST,
+            method: f.mutability.into(),
             inner: Arc::from(index) as Arc<RawHandler>,
         }
     }
@@ -474,28 +508,116 @@ impl From<AllowOrigin> for Cors {
     }
 }
 
-#[test]
-fn allow_origin_from_str() {
-    fn check(text: &str, expected: AllowOrigin) {
-        let from_str = AllowOrigin::from_str(text).unwrap();
-        assert_eq!(from_str, expected);
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn allow_origin_from_str() {
+        fn check(text: &str, expected: AllowOrigin) {
+            let from_str = AllowOrigin::from_str(text).unwrap();
+            assert_eq!(from_str, expected);
+        }
+
+        check(r#"*"#, AllowOrigin::Any);
+        check(
+            r#"http://example.com"#,
+            AllowOrigin::Whitelist(vec!["http://example.com".to_string()]),
+        );
+        check(
+            r#"http://a.org, http://b.org"#,
+            AllowOrigin::Whitelist(vec!["http://a.org".to_string(), "http://b.org".to_string()]),
+        );
+        check(
+            r#"http://a.org, http://b.org, "#,
+            AllowOrigin::Whitelist(vec!["http://a.org".to_string(), "http://b.org".to_string()]),
+        );
+        check(
+            r#"http://a.org,http://b.org"#,
+            AllowOrigin::Whitelist(vec!["http://a.org".to_string(), "http://b.org".to_string()]),
+        );
     }
 
-    check(r#"*"#, AllowOrigin::Any);
-    check(
-        r#"http://example.com"#,
-        AllowOrigin::Whitelist(vec!["http://example.com".to_string()]),
-    );
-    check(
-        r#"http://a.org, http://b.org"#,
-        AllowOrigin::Whitelist(vec!["http://a.org".to_string(), "http://b.org".to_string()]),
-    );
-    check(
-        r#"http://a.org, http://b.org, "#,
-        AllowOrigin::Whitelist(vec!["http://a.org".to_string(), "http://b.org".to_string()]),
-    );
-    check(
-        r#"http://a.org,http://b.org"#,
-        AllowOrigin::Whitelist(vec!["http://a.org".to_string(), "http://b.org".to_string()]),
-    );
+    fn assert_responses_eq(left: HttpResponse, right: HttpResponse) {
+        assert_eq!(left.status(), right.status());
+        assert_eq!(left.headers(), right.headers());
+        assert_eq!(left.body(), right.body());
+    }
+
+    #[test]
+    fn test_create_warning_header() {
+        assert_eq!(
+            &create_warning_header("Description"),
+            "299 - \"Description\""
+        );
+    }
+
+    #[test]
+    fn json_responses() {
+        use chrono::TimeZone;
+
+        let actual_response = json_response(Actuality::Actual, 123);
+        assert_responses_eq(actual_response, HttpResponse::Ok().json(123));
+
+        let deprecated_response_no_deadline = json_response(
+            Actuality::Deprecated {
+                discontinued_on: None,
+                description: None,
+            },
+            123,
+        );
+        let expected_warning_text =
+            "Deprecated API: This endpoint is deprecated, \
+             see the service documentation to find an alternative. \
+             Currently there is no specific date for disabling this endpoint.";
+        let expected_warning = create_warning_header(expected_warning_text);
+        assert_responses_eq(
+            deprecated_response_no_deadline,
+            HttpResponse::Ok()
+                .header(header::WARNING, expected_warning)
+                .json(123),
+        );
+
+        let description = "Docs can be found on docs.rs".to_owned();
+        let deprecated_response_with_description = json_response(
+            Actuality::Deprecated {
+                discontinued_on: None,
+                description: Some(description),
+            },
+            123,
+        );
+        let expected_warning_text =
+            "Deprecated API: This endpoint is deprecated, \
+             see the service documentation to find an alternative. \
+             Currently there is no specific date for disabling this endpoint. \
+             Additional information: Docs can be found on docs.rs.";
+        let expected_warning = create_warning_header(expected_warning_text);
+        assert_responses_eq(
+            deprecated_response_with_description,
+            HttpResponse::Ok()
+                .header(header::WARNING, expected_warning)
+                .json(123),
+        );
+
+        let deadline = chrono::Utc.ymd(2020, 12, 31).and_hms(23, 59, 59);
+
+        let deprecated_response_deadline = json_response(
+            Actuality::Deprecated {
+                discontinued_on: Some(deadline),
+                description: None,
+            },
+            123,
+        );
+        let expected_warning_text =
+            "Deprecated API: This endpoint is deprecated, \
+             see the service documentation to find an alternative. \
+             The old API is maintained until Thu, 31 Dec 2020 23:59:59 GMT.";
+        let expected_warning = create_warning_header(expected_warning_text);
+        assert_responses_eq(
+            deprecated_response_deadline,
+            HttpResponse::Ok()
+                .header(header::WARNING, expected_warning)
+                .json(123),
+        );
+    }
 }