+++ test-suite/testkit/src/lib.rs
@@ -100,8 +100,8 @@
 //!
 //!     // Check results with api.
 //!     let api = testkit.api();
-//!     let explorer_api = api.public(ApiKind::Explorer);
-//!     let response: BlocksRange = explorer_api
+//!     let response: BlocksRange = api
+//!         .public(ApiKind::Explorer)
 //!         .query(&BlocksQuery {
 //!             count: 10,
 //!             ..Default::default()
@@ -113,7 +113,8 @@
 //!     assert_eq!(range.start, Height(0));
 //!     assert_eq!(range.end, Height(3));
 //!
-//!     let info = explorer_api
+//!     let info = api
+//!         .public(ApiKind::Explorer)
 //!         .query(&TransactionQuery::new(tx1.object_hash()))
 //!         .get::<TransactionInfo>("v1/transactions")
 //!         .unwrap();
@@ -466,7 +467,7 @@ impl TestKit {
 
         let guard = self.processing_lock.lock().unwrap();
         let (block_hash, patch) = self.blockchain.create_patch(
-            validator_id,
+            validator_id.into(),
             new_block_height,
             tx_hashes,
             &mut BTreeMap::new(),